[Gold V] Title: 행복 유치원, Time: 56 ms, Memory: 8296 KB -BaekjoonHub

+++ 백준/Gold/13164. 행복 유치원/행복 유치원.cc
@@ -7,29 +7,29 @@ using namespace std;
 int main()
 {
     ios_base::sync_with_stdio(false); cin.tie(NULL); cout.tie(NULL);
-    long int N, K;
-    vector<long int> v;
+    int N, K;
+    vector<int> v;
     cin >> N >> K;
 
     for (int i = 0; i < N; i++)
     {
-        long int num;
+        int num;
         cin >> num;
         v.push_back(num);
     }
 
     sort(v.begin(), v.end());
 
-    vector<long int> diff;
+    vector<int> diff;
     for (int i = 0; i < v.size() - 1; i++)
     {
-        long int result = v[i + 1] - v[i];
+        int result = v[i + 1] - v[i];
         diff.push_back(result);
     }
 
     sort(diff.begin(), diff.end());
 
-    long int answer = 0;
+    int answer = 0;
     for (int i = 0; i < diff.size()-(K-1); i++)
     {
         answer += diff[i];
